Migrated over from c GPIO subscibe tools to C++ EXTISubscribe class.
