+++ Firmware/MotorControl/encoder.cpp
@@ -59,15 +59,15 @@ void Encoder::enc_index_cb() {
     }
 
     // Disable interrupt
-    GPIO_unsubscribe(hw_config_.index_port, hw_config_.index_pin);
+    SubscribeEXTI.unsubscribe(hw_config_.index_port, hw_config_.index_pin);
 }
 
 void Encoder::set_idx_subscribe(bool override_enable) {
     if (config_.use_index && (override_enable || !config_.find_idx_on_lockin_only)) {
-        GPIO_subscribe(hw_config_.index_port, hw_config_.index_pin, GPIO_PULLDOWN,
+        SubscribeEXTI.subscribe(hw_config_.index_port, hw_config_.index_pin, GPIO_PULLDOWN,
                 enc_index_cb_wrapper, this);
     } else if (!config_.use_index || config_.find_idx_on_lockin_only) {
-        GPIO_unsubscribe(hw_config_.index_port, hw_config_.index_pin);
+        SubscribeEXTI.unsubscribe(hw_config_.index_port, hw_config_.index_pin);
     }
 }
